식당 리뷰 구현 - 식당 리뷰 리스트 가져오기
# 식당 상세화면에서 리뷰 보여주기

+++ app/src/main/java/kr/co/bepo/fooddeliveryapp/widget/adapter/viewholder/review/ReviewViewHolder.kt
@@ -32,7 +32,7 @@ class ReviewViewHolder(
             reviewThumbnailImageView.toGone()
         }
         reviewUserTextView.text = model.title
-        ratingBar.rating = model.grade.toFloat()
+        ratingBar.rating = model.grade
         reviewTextView.text = model.description
     }
